Rename ItemUiState to SegmentsItemUiState
